feat: API 형태 추가
