+++ src/main/java/in/koreatech/koin/domain/user/service/UserService.java
@@ -10,7 +10,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import in.koreatech.koin.domain.user.dto.EmailCheckExistsRequest;
 import in.koreatech.koin.domain.user.dto.NicknameCheckExistsRequest;
-import in.koreatech.koin.domain.user.dto.NotificationStatusResponse;
 import in.koreatech.koin.domain.user.dto.UserLoginRequest;
 import in.koreatech.koin.domain.user.dto.UserLoginResponse;
 import in.koreatech.koin.domain.user.dto.UserTokenRefreshRequest;
@@ -92,22 +91,5 @@ public class UserService {
         userRepository.findByNickname(request.nickname()).ifPresent(user -> {
             throw DuplicationEmailException.withDetail("nickname: " + request.nickname());
         });
-    }
-
-    public NotificationStatusResponse checkNotification(Long userId) {
-        User user = userRepository.getById(userId);
-        return new NotificationStatusResponse(user.getDeviceToken() != null);
-    }
-
-    @Transactional
-    public void permitNotification(Long userId, String deviceToken) {
-        User user = userRepository.getById(userId);
-        user.permitNotification(deviceToken);
-    }
-
-    @Transactional
-    public void rejectNotification(Long userId) {
-        User user = userRepository.getById(userId);
-        user.rejectNotification();
     }
 }
