+++ src/main/java/in/koreatech/koin/global/domain/notification/NotificationService.java
@@ -1,7 +1,11 @@
 package in.koreatech.koin.global.domain.notification;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import in.koreatech.koin.domain.user.dto.NotificationStatusResponse;
+import in.koreatech.koin.domain.user.model.User;
+import in.koreatech.koin.domain.user.repository.UserRepository;
 import in.koreatech.koin.global.fcm.FcmClient;
 import lombok.RequiredArgsConstructor;
 
@@ -9,6 +13,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class NotificationService {
 
+    private final UserRepository userRepository;
     private final NotificationRepository notificationRepository;
     private final FcmClient fcmClient;
 
@@ -23,5 +28,22 @@ public class NotificationService {
             notification.getUrl(),
             notification.getType()
         );
+    }
+
+    public NotificationStatusResponse checkNotification(Long userId) {
+        User user = userRepository.getById(userId);
+        return new NotificationStatusResponse(user.getDeviceToken() != null);
+    }
+
+    @Transactional
+    public void permitNotification(Long userId, String deviceToken) {
+        User user = userRepository.getById(userId);
+        user.permitNotification(deviceToken);
+    }
+
+    @Transactional
+    public void rejectNotification(Long userId) {
+        User user = userRepository.getById(userId);
+        user.rejectNotification();
     }
 }
