+++ src/main/java/in/koreatech/koin/global/domain/notification/controller/NotificationApi.java
@@ -1,0 +1,71 @@
+package in.koreatech.koin.global.domain.notification.controller;
+
+import static in.koreatech.koin.domain.user.model.UserType.*;
+import static in.koreatech.koin.domain.user.model.UserType.COOP;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+
+import in.koreatech.koin.domain.user.dto.NotificationPermitRequest;
+import in.koreatech.koin.domain.user.dto.NotificationStatusResponse;
+import in.koreatech.koin.global.auth.Auth;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
+
+@Tag(name = "(Normal) Notification: 알림", description = "알림 관련 API")
+public interface NotificationApi {
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "200"),
+            @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
+            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
+        }
+    )
+    @Operation(summary = "푸쉬알림 동의 여부 조회")
+    @GetMapping("/notification")
+    ResponseEntity<NotificationStatusResponse> checkNotificationStatus(
+        @Auth(permit = {STUDENT, OWNER, COOP}) Long memberId
+    );
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "200"),
+            @ApiResponse(responseCode = "400"),
+            @ApiResponse(responseCode = "401"),
+            @ApiResponse(responseCode = "403"),
+            @ApiResponse(responseCode = "404"),
+        }
+    )
+    @Operation(summary = "푸쉬알림 동의")
+    @PostMapping("/notification")
+    ResponseEntity<Void> permitNotification(
+        @Auth(permit = {STUDENT, OWNER, COOP}) Long memberId,
+        @Valid @RequestBody NotificationPermitRequest request
+    );
+
+    @ApiResponses(
+        value = {
+            @ApiResponse(responseCode = "200"),
+            @ApiResponse(responseCode = "400"),
+            @ApiResponse(responseCode = "401"),
+            @ApiResponse(responseCode = "403"),
+            @ApiResponse(responseCode = "404"),
+        }
+    )
+    @Operation(summary = "푸쉬알림 거절")
+    @DeleteMapping("/notification")
+    ResponseEntity<Void> rejectNotification(
+        @Auth(permit = {STUDENT, OWNER, COOP}) Long memberId
+    );
+}
